requirements.txt
# Core dependencies
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Bootstrap==3.3.7.1
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
asyncpg==0.28.0
alembic==1.12.0

# Asynchronous support
aiohttp==3.8.5
asyncio==3.4.3
uvloop==0.17.0; sys_platform != 'win32'

# Data processing
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1
openpyxl==3.1.2

# Caching and messaging
redis==4.6.0
python-json-logger==2.0.7

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0
pydantic==2.4.2

# Monitoring and metrics
prometheus-client==0.17.1
psutil==5.9.5

# Utilities
tqdm==4.66.1
tenacity==8.2.3
dateparser==1.1.8
backoff==2.2.1
humanize==4.8.0
cachetools==5.3.1

# Security
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1

# Development and testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.5.1

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
